+++ src/main/java/com/pmv/service/CpuService.java
@@ -6,13 +6,13 @@ import com.pmv.entity.Cpu;
 
 public interface CpuService {
 	
-	public abstract List<Cpu> getAll();
+	public abstract  List<Cpu> getAll();
 	
 	public abstract Cpu getOne(Long cpuId);
 	
 	public abstract Cpu addOne(Cpu cpu);
 	
-	public abstract void delete(int cpuId);
+	public abstract void delete(Long cpuId);
 	
 	public abstract boolean exists(Long cpuId);
